refactor authorization

+++ StockMarket.Api/Controllers/PortfolioController.cs
@@ -17,6 +17,7 @@ public class PortfolioController : ControllerBase
     }
 
     [HttpGet]
+    [Authorize]
     public async Task<ActionResult> GetUserPortfolio()
     {
         var userName = User.Claims.FirstOrDefault(claim => claim.Type == ClaimTypes.Email)!.Value;
@@ -26,13 +27,21 @@ public class PortfolioController : ControllerBase
     }
 
     [HttpPost]
+    [Authorize]
     public async Task<ActionResult> CreatePortfolio(CreatePortfolioDto model)
     {
         if (!ModelState.IsValid)
         {
             return BadRequest(ModelState);
         }
-        var result = await _portfolioService.CreatePortfolioAsync(model);
+
+		var userId = User.Claims.FirstOrDefault(u => u.Type == ClaimTypes.NameIdentifier)?.Value ?? "NA";
+
+		if (userId == "NA") return BadRequest("user not found..!!!");
+
+		model.UserId = userId;
+
+		var result = await _portfolioService.CreatePortfolioAsync(model);
         if (!result.IsSuccessed)
         {
             return BadRequest(result);
@@ -42,12 +51,20 @@ public class PortfolioController : ControllerBase
 	}
 
     [HttpDelete]
+    [Authorize]
     public async Task<ActionResult> DeletePortfolio(DeletePortfolioDto model)
     {
         if(!ModelState.IsValid)
         {
             return BadRequest(ModelState);
         }
+
+		var userId = User.Claims.FirstOrDefault(u => u.Type == ClaimTypes.NameIdentifier)?.Value ?? "NA";
+
+		if (userId == "NA") return BadRequest("user not found..!!!");
+
+		model.UserId = userId;
+
 		var result = await _portfolioService.DeletePortfolioAsync(model);
 		if (!result.IsSuccessed)
 		{
